23443: Update PreSubmitInfo component for use with 1990S and 0994
