+++ src/applications/edu-benefits/1990s/containers/PreSubmitInfo.jsx
@@ -1,70 +1,13 @@
-import React, { useEffect } from 'react';
-import Checkbox from '@department-of-veterans-affairs/component-library/Checkbox';
+import React from 'react';
 import { connect } from 'react-redux';
 
 // platform - form-system actions
 import { setPreSubmit as setPreSubmitAction } from 'platform/forms-system/src/js/actions';
 import RadioButtons from '@department-of-veterans-affairs/component-library/RadioButtons';
+import PreSubmitInfo from '../../containers/PreSubmitInfo';
 
-function PreSubmitNotice({
-  formData,
-  showError,
-  onSectionComplete,
-  setPreSubmit,
-}) {
-  const privacyAgreementAccepted = formData.privacyAgreementAccepted || false;
+function PreSubmitNotice({ formData, onSectionComplete, setPreSubmit }) {
   const vrrapConfirmation = formData.vrrapConfirmation;
-
-  // set AGREED (onSectionComplete) to value of privacyAgreementAccepted
-  // if goes to another page (unmount), set AGREED (onSectionComplete) to false
-  useEffect(
-    () => {
-      onSectionComplete(privacyAgreementAccepted);
-
-      return () => {
-        onSectionComplete(false);
-      };
-    },
-
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    [privacyAgreementAccepted],
-  );
-
-  const privacyAgreementLabel = (
-    <span>
-      I have read and accept the{' '}
-      <a
-        aria-label="Privacy policy, will open in new tab"
-        target="_blank"
-        href="/privacy-policy/"
-      >
-        privacy policy
-      </a>
-    </span>
-  );
-
-  const privacyAgreement = (
-    <div>
-      <div>
-        <strong>Note:</strong> According to federal law, there are criminal
-        penalties, including a fine and/or imprisonment for up to 5 years, for
-        withholding information or for providing incorrect information. (See 18
-        U.S.C. 1001)
-      </div>
-      <Checkbox
-        required
-        checked={privacyAgreementAccepted}
-        onValueChange={value => setPreSubmit('privacyAgreementAccepted', value)}
-        name={'privacyAgreementAccepted'}
-        errorMessage={
-          showError && !privacyAgreementAccepted
-            ? 'You must accept the privacy policy before continuing.'
-            : undefined
-        }
-        label={privacyAgreementLabel}
-      />
-    </div>
-  );
 
   const confirmEligibilityNote = (
     <div>
@@ -113,7 +56,10 @@ function PreSubmitNotice({
   return (
     <>
       {confirmEligibilityNote}
-      {privacyAgreement}
+      <PreSubmitInfo
+        formData={formData}
+        onSectionComplete={onSectionComplete}
+      />
     </>
   );
 }
